Added initial Frontend

+++ backend/controllers/goalControllers.js
@@ -1,12 +1,14 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,7 +22,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -36,6 +39,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    //Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not Authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -44,6 +61,7 @@ const updateGoal = asyncHandler(async (req, res) => {
 
     res.status(200).json(updatedGoal)
 })
+
 // @desc Delete Goal
 // @route DELETE /api/goals/:id
 // @access Private
@@ -55,7 +73,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
 
+    //Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not Authorized')
+    }
+    
     await goal.deleteOne()
 
     res.status(200).json({id: req.params.id})
